+++ octodns/processor/meta.py
@@ -111,13 +111,13 @@ class MetaProcessor(BaseProcessor):
             self.record_name,
             {'ttl': self.ttl, 'type': 'TXT', 'values': self.values},
             # we may be passing in empty values here to be filled out later in
-            # process_target_zone
+            # process_source_and_target_zones
             lenient=True,
         )
         desired.add_record(meta)
         return desired
 
-    def process_target_zone(self, existing, target):
+    def process_source_and_target_zones(self, desired, existing, target):
         if self.include_provider:
             # look for the meta record
             for record in sorted(existing.records):
@@ -129,7 +129,7 @@ class MetaProcessor(BaseProcessor):
                     existing.add_record(record, replace=True)
                     break
 
-        return existing
+        return desired, existing
 
     def _up_to_date(self, change):
         # existing state, if there is one