+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Record.lenient property added similar to other common/standard _octodns data
 * Processor.process_source_and_target_zones added to support modifying both the
   desired and/or existing zones just prior to computing changes.
+* Fix an issue in MetaProcessor/Manager.include_meta where include_provider
+  wasn't correctly taking effect
 
 ## v1.3.0 - 2023-11-14 - New and improved processors