+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   desired and/or existing zones just prior to computing changes.
 * Fix an issue in MetaProcessor/Manager.include_meta where include_provider
   wasn't correctly taking effect
+* Fix bug with Record.copy when values is an empty list []
 
 ## v1.3.0 - 2023-11-14 - New and improved processors